Modification
I needed this code for a job, and I made some changes, making it shorter and easier to understand. Giving that the code is yours it was only fair to give it to you. Im sorry it is in spanish, but you should be able to understand it no problem nonetheless.